Movement, Sound and Level loading
Added the ability to give the rocket thrust and to rotate. Also added a sound effect that only plays when the player is giving the rocket thrust. Lastly, added tags, another scene and the logic to reload if the player crashes or load the next level if they reaching the landing pad. The levels will cycle back to the beginning if you complete the final level.