Sound effects, rocket design, success and crash framework
Added new sound effects that play when you crash or reach the goal. Logic was added to ensure that the appropriate sound effects are playing at the right time. This was primarily accomplished by adding a state boolean and two new sequence handling methods to track when the game is loading and turning off events triggered by collisions (sound effects and movement inputs). Lastly, re-did the rocket design to make it look more like a rocket and added a base to it to make it easier to land.