Lights, 'Over' make over and moving obstacles
Added point lights to the rocket to make it stand out more and to have the radius of the lights act as an early warning system when getting to close to obstacles. Added a glow effect to the launch and landing pads through the emission characteristic on their material. Did a visual redesign of the 'Over' level changing up the colour scheme and making it darker to give it more of a space feel. Also, added lights to 'Over' to mark places the player needs to be careful of. Lastly, wrote a script that makes obstacles move around the screen using the sin function.